services/backend/requirements.txt
fastapi==0.110.0
uvicorn[standard]==0.27.1
sqlalchemy==2.0.29
psycopg2-binary==2.9.9
alembic==1.13.1
passlib[bcrypt]==1.7.4
bcrypt==3.2.2
python-jose[cryptography]==3.3.0
pydantic==2.6.4
pydantic-settings==2.2.1
pgvector==0.2.5
scikit-learn==1.4.1.post1
numpy==1.26.4
openai==1.30.1
anthropic==0.25.2
sentence-transformers==2.6.1
langdetect==1.0.9
rapidfuzz==3.6.2
beautifulsoup4==4.12.3
feedparser==6.0.11
httpx==0.27.0
python-dateutil==2.9.0.post0
orjson==3.10.3
email-validator==2.1.1
pyyaml==6.0.1
trafilatura==1.8.1
lxml==5.1.0
